% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_bc.R
\name{esm_bc}
\alias{esm_bc}
\title{Wrapper function to create barchart for each date / measurement.}
\usage{
esm_bc(data = NULL, var_date = NULL, vars_meas = NULL, bars = NULL,
  outcome = NULL, vis_options = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{var_date}{A string with the name of the date-variable for the different facets.}

\item{outcome}{A string with name of variable specifying the outcome.}

\item{vis_options}{A string with name of variable specifying the different groups which will be represented as different lines
The possibilities for the vis_options are:
"axis_limits = ..."; vector with lower and upper limit of y-axis (e.g., c(0, 10))}

\item{var_meas}{A vector with strings of variable names that determines order of x-axis.}

\item{lines}{A string with name of variable specifying the different groups which will be represented as different lines.}
}
\value{
A ggplot-object/graph.
}
\description{
Wrapper function to create barchart for each date / measurement.
}
