% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_ts.R
\name{esm_ts}
\alias{esm_ts}
\title{Wrapper function to create timeseries line plot.}
\usage{
esm_ts(data = NULL, var_date = NULL, lines = NULL, outcome = NULL,
  vis_options = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{var_date}{A string with the name of the date-variable for the x-axis.}

\item{lines}{A string with name of variable specifying the different groups which will be represented as different lines.}

\item{outcome}{A string with name of variable specifying the outcome.}

\item{vis_options}{A list with options for the graph.
The possibilities for the vis_options are:
"line = TRUE/FALSE"; whether or not a line should be drawn
"point = TRUE/FALSE"; whether or not points/dots should be drawn
"smooth = TRUE/FALSE"; whether or not smoothening line should be drawn.
"kernel = ..."; parameter to control the degree of smoothening (between 0 and 1).
"se_band = TRUE/FALSE"; whether or not confidence band should be drawn around smoothening line.
"axis_limits = ..."; vector with lower and upper limit of y-axis (e.g., c(0, 10))}
}
\value{
A ggplot-object/graph.
}
\description{
Wrapper function to create timeseries line plot.
}
