% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_vis.R
\name{esm_vis}
\alias{esm_vis}
\title{Wrapper function to create desired visualisation.}
\usage{
esm_vis(data = NULL, var_date = NULL, format_date = NULL,
  vars_meas = NULL, vars_groups = NULL, vars_event = NULL,
  vars_descr = NULL, ID = NULL, type_vis = "timeseries",
  vis_options = NULL, time_frame = "all", sel_period = NULL,
  sel_period_zoom = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{var_date}{A string with the name of the date-variable.}

\item{format_date}{A string with the format of the date (e.g., "ymd", "ymd_HM").}

\item{vars_meas}{A vector of variable names that need to be visualised.}

\item{vars_groups}{A vector of specific variable names that are in vars_meas.}

\item{vars_event}{A vector of names of variables that describe events.}

\item{vars_descr}{A vector of names of variables that describe events.}

\item{ID}{A list that must contain the elements ID_var and ID; ID_var must be a string with a variable name, and ID must be a string of the unique identifier.}

\item{type_vis}{The type of visualisation required; the options are "timeseries" (default), "zoom", "barchart", and "network".}

\item{vis_options}{A list with visualisation options. See the functions esm_ts, esm_bc, and esm_nw for further help.}

\item{time_frame}{A vector with the first and last measurement.}

\item{sel_period}{A vector with the first and last measurement for selecting a specific period.}

\item{sel_period_zoom}{A vector with the first and last measurement for selecting a specific period in the "zoom" graph.}
}
\value{
A ggplot-object/graph.
}
\description{
Wrapper function to create desired visualisation.
}
