% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm_zoom.R
\name{esm_zoom}
\alias{esm_zoom}
\title{Wrapper function to create timeseries line plot and below a zoomed-in plot.}
\usage{
esm_zoom(data = NULL, data_zoom = NULL, var_date = NULL, lines = NULL,
  outcome = NULL, vis_options = NULL, vars_events = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{data_zoom}{A dataframe with the period selected for zoom.}

\item{var_date}{A string with the name of the date-variable for the x-axis.}

\item{lines}{A string with name of variable specifying the different groups which will be represented as different lines.}

\item{outcome}{A string with name of variable specifying the outcome.}

\item{vis_options}{A list with options for the graph.
The possibilities for the vis_options are:
"line = TRUE/FALSE"; whether or not a line should be drawn
"point = TRUE/FALSE"; whether or not points/dots should be drawn
"smooth = TRUE/FALSE"; whether or not smoothening line should be drawn.
"kernel = ..."; parameter to control the degree of smoothening (between 0 and 1).
"se_band = TRUE/FALSE"; whether or not confidence band should be drawn around smoothening line.
"axis_limits = ..."; vector with lower and upper limit of y-axis (e.g., c(0, 10))}

\item{vars_events}{A vector with XXXXX.}
}
\value{
A ggplot-object/graph.
}
\description{
Wrapper function to create timeseries line plot and below a zoomed-in plot.
}
